hw9-final-project/requirements.txt
Flask==2.3.2
google-cloud-datastore==2.15.0
json2html==1.3.0
email-validator==2.0.0.post2
python-jose
flask-cors
six
python-dotenv
requests
authlib
protobuf==3.20.*
